+++ Moco/Moco/MocoGoal/MocoTranslationTrackingGoal.cpp
@@ -162,7 +162,7 @@ void MocoTranslationTrackingGoal::printDescriptionImpl(std::ostream& stream) con
     stream << "        ";
     stream << "translation reference file: "
            << get_translation_reference_file() << std::endl;
-    for (int i = 0; i < m_frame_paths.size(); i++) {
+    for (int i = 0; i < (int)m_frame_paths.size(); i++) {
         stream << "        ";
         stream << "frame " << i << ": " << m_frame_paths[i] << ", ";
         stream << "weight: " << m_translation_weights[i] << std::endl;
